use statefull widget for series thumbnail
