+++ lib/screen/feralfile_home/artwork_view.dart
@@ -1,23 +1,15 @@
 import 'dart:async';
 
 import 'package:autonomy_flutter/common/injector.dart';
-import 'package:autonomy_flutter/model/ff_exhibition.dart';
 import 'package:autonomy_flutter/model/ff_list_response.dart';
 import 'package:autonomy_flutter/model/ff_series.dart';
-import 'package:autonomy_flutter/screen/app_router.dart';
-import 'package:autonomy_flutter/screen/exhibition_details/exhibition_detail_page.dart';
-import 'package:autonomy_flutter/screen/feralfile_artwork_preview/feralfile_artwork_preview_page.dart';
 import 'package:autonomy_flutter/screen/feralfile_home/filter_bar.dart';
-import 'package:autonomy_flutter/screen/feralfile_series/feralfile_series_page.dart';
 import 'package:autonomy_flutter/service/feralfile_service.dart';
-import 'package:autonomy_flutter/util/feralfile_alumni_ext.dart';
 import 'package:autonomy_flutter/util/log.dart';
-import 'package:autonomy_flutter/util/series_ext.dart';
-import 'package:autonomy_flutter/view/feralfile_cache_network_image.dart';
+import 'package:autonomy_flutter/view/ff_series_tappable_thumbnail.dart';
 import 'package:autonomy_flutter/view/loading.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:feralfile_app_theme/feral_file_app_theme.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 class ExploreSeriesView extends StatefulWidget {
@@ -193,7 +185,6 @@ class SeriesView extends StatefulWidget {
 
 class _SeriesViewState extends State<SeriesView> {
   late ScrollController _scrollController;
-  String? _navigatingSeriesId;
 
   @override
   void initState() {
@@ -247,135 +238,14 @@ class _SeriesViewState extends State<SeriesView> {
         ],
       );
 
-  Widget _seriesInfo(BuildContext context, FFSeries series) {
-    final theme = Theme.of(context);
-    final defaultStyle = theme.textTheme.ppMori400White12;
-    return Row(
-      children: [
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                series.artist?.displayAlias ?? '',
-                style: defaultStyle,
-                overflow: TextOverflow.ellipsis,
-              ),
-              const SizedBox(height: 4),
-              Text(
-                series.displayTitle,
-                style: defaultStyle.copyWith(
-                  fontWeight: FontWeight.bold,
-                  fontStyle: FontStyle.italic,
-                ),
-                overflow: TextOverflow.ellipsis,
-              ),
-              if (series.exhibition != null) ...[
-                const SizedBox(height: 12),
-                RichText(
-                  textScaler: MediaQuery.textScalerOf(context),
-                  text: TextSpan(
-                    style: defaultStyle.copyWith(color: AppColor.auQuickSilver),
-                    children: [
-                      TextSpan(
-                        text: 'exhibited_in'.tr(),
-                      ),
-                      TextSpan(
-                        text: series.exhibition!.title,
-                        style: const TextStyle(
-                          decoration: TextDecoration.underline,
-                        ),
-                        recognizer: TapGestureRecognizer()
-                          ..onTap = () {
-                            _gotoExhibitionDetails(context, series.exhibition!);
-                          },
-                      ),
-                    ],
-                  ),
-                ),
-              ]
-            ],
-          ),
-        )
-      ],
-    );
-  }
-
   Widget _seriesItem(BuildContext context, FFSeries series, Border border) =>
-      GestureDetector(
-        onTap: () async {
-          await _gotoSeriesDetails(context, series);
-        },
-        child: Container(
-          decoration: BoxDecoration(
-            // border on the top and right
-            border: border,
-            color: Colors.transparent,
-          ),
-          padding: const EdgeInsets.all(8),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Expanded(
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Expanded(
-                      child: _navigatingSeriesId == series.id
-                          ? const LoadingWidget()
-                          : FFCacheNetworkImage(
-                              imageUrl: series.thumbnailUrl ?? '',
-                              fit: BoxFit.fitWidth,
-                            ),
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 8),
-              _seriesInfo(context, series),
-            ],
-          ),
+      Container(
+        decoration: BoxDecoration(
+          // border on the top and right
+          border: border,
+          color: Colors.transparent,
         ),
+        padding: const EdgeInsets.all(8),
+        child: FfSeriesInfoThumbnail(series: series),
       );
-
-  Future<void> _gotoSeriesDetails(BuildContext context, FFSeries series) async {
-    if (series.isSingle) {
-      setState(() {
-        _navigatingSeriesId = series.id;
-      });
-      final artwork =
-          await injector<FeralFileService>().getFirstViewableArtwork(series.id);
-      if (artwork != null) {
-        if (context.mounted) {
-          unawaited(Navigator.of(context).pushNamed(
-            AppRouter.ffArtworkPreviewPage,
-            arguments: FeralFileArtworkPreviewPagePayload(
-              artwork: artwork.copyWith(series: series),
-            ),
-          ));
-        }
-      }
-      if (context.mounted) {
-        setState(() {
-          _navigatingSeriesId = null;
-        });
-      }
-    } else {
-      unawaited(Navigator.of(context).pushNamed(
-        AppRouter.feralFileSeriesPage,
-        arguments: FeralFileSeriesPagePayload(
-          seriesId: series.id,
-          exhibitionId: series.exhibitionID,
-        ),
-      ));
-    }
-  }
-
-  void _gotoExhibitionDetails(BuildContext context, Exhibition exhibition) {
-    unawaited(Navigator.of(context).pushNamed(AppRouter.exhibitionDetailPage,
-        arguments: ExhibitionDetailPayload(
-          exhibitions: [exhibition],
-          index: 0,
-        )));
-  }
 }
